api/pyproject.toml
[tool.poetry]
name = "epidemic"
version = "0.1.0"
description = "Epidemic Sound API Server."
authors = ["Moeen Zamani <moeen.zamani@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.10"
Django = "4.0.1"
djangorestframework = "3.13.1"
django-cors-headers = "3.11.0"
requests = "2.27.1"
psycopg2-binary = "2.9.3"
dj-database-url = "0.5.0"
pytest-cov = "3.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "1.6.0"
ipython = "8.0.1"
pytest = "6.2.5"
pytest-django = "4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
